TransferService
*

Creating and implementing basic methods for transferService

*

+++ src/main/java/br/com/vsgdev/emotionsRoute/repository/TransferRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.vsgdev.emotionsRoute.model.Transfer;
 
 @Repository
-public interface TransferRepository extends CrudRepository<Transfer, Integer> {
+public interface TransferRepository extends CrudRepository<Transfer, Long> {
 
 }
